Program and tests refactoring

+++ hw1/HomeWork1/Calculator.cs
@@ -2,7 +2,7 @@
 {
     public static class Calculator
     {
-        public static int Calculate(int val1, string operation, int val2)
+        public static double Calculate(double val1, string operation, double val2)
         {
             var result = operation switch
             {
